requirements.txt
dash>=2.14
dash-bootstrap-components>=1.5
plotly>=5.18
pandas>=2.0
numpy>=1.24
scipy>=1.11
requests-oauthlib>=1.3
python-dotenv>=1.0
keyring>=24.0
